Switching from osgUtil::IntersectVisitor to osgUtil::LineSegmentIntersector.  LineSegmentIntersector gives the intersection point back in double precision.

+++ trunk/Usul/Interfaces/IScreenCapture.h
@@ -12,7 +12,7 @@
 
 #include "Usul/Interfaces/IUnknown.h"
 
-namespace osg { class Image; class Vec3f; class Quat; }
+namespace osg { class Image; class Vec3d; class Quat; }
 
 namespace Usul {
 namespace Interfaces {
@@ -26,7 +26,7 @@ struct IScreenCapture : public Usul::Interfaces::IUnknown
   /// Id for this interface.
   enum { IID = 1275760u };
 
-  virtual osg::Image* screenCapture ( const osg::Vec3f& center, float distance, const osg::Quat& rotation, unsigned int height, unsigned int width ) const = 0;
+  virtual osg::Image* screenCapture ( const osg::Vec3d& center, double distance, const osg::Quat& rotation, unsigned int height, unsigned int width ) const = 0;
   virtual osg::Image* screenCapture ( unsigned int height, unsigned int width ) const = 0;
 
 }; // struct IScreenCapture
